update
Added a new feature to enemy spawner script. A door that takes you to whatever scene you want, and it appears when waves are over. However, whenever the waves stop, we seem to be getting errors.

Also removed the ground for now and the lava (lava was made prefab, and might be used in actual map)